project list is a listview now
