+++ src/main/java/de/doubleslash/keeptime/controller/Controller.java
@@ -289,4 +289,9 @@ public class Controller {
 
       return seconds;
    }
+
+   public long getCurrentWorkSeconds() {
+      return Duration.between(model.activeWorkItem.get().getStartTime(), model.activeWorkItem.get().getEndTime())
+            .getSeconds();
+   }
 }
