+++ src/main/resources/layouts/ViewLayoutPopup.fxml
@@ -16,20 +16,18 @@
  You should have received a copy of the GNU General Public License
  along with this program. If not, see <http://www.gnu.org/licenses/>. -->
 
-<?import java.lang.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.text.*?>
 <?import javafx.scene.control.ListView?>
 <?import javafx.scene.control.TextField?>
 <?import javafx.scene.layout.VBox?>
+<?import javafx.scene.text.Font?>
 
-<VBox prefWidth="100.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="de.doubleslash.keeptime.viewpopup.ViewControllerPopup">
+<VBox prefWidth="100.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="de.doubleslash.keeptime.viewpopup.ViewControllerPopup">
    <children>
       <TextField fx:id="searchTextField" promptText="Search">
          <font>
             <Font name="Open Sans Regular" size="12.0" />
-         </font></TextField>
-      <ListView fx:id="projectListView" prefWidth="100.0" />
+         </font>
+      </TextField>
+      <ListView fx:id="projectListView" prefWidth="100.0" stylesheets="@../css/application.css" />
    </children>
 </VBox>
